+++ tests/fkooman/Rest/Plugin/Authentication/Bearer/BearerAuthenticationTest.php
@@ -22,6 +22,7 @@ require_once __DIR__.'/Test/TestValidator.php';
 use fkooman\Http\Request;
 use PHPUnit_Framework_TestCase;
 use fkooman\Rest\Plugin\Authentication\Bearer\Test\TestValidator;
+use fkooman\Http\Exception\UnauthorizedException;
 
 class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
 {
@@ -36,6 +37,28 @@ class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
                 'SCRIPT_NAME' => '/index.php',
                 'REQUEST_METHOD' => 'GET',
                 'HTTP_AUTHORIZATION' => 'Bearer t_fkooman',
+            )
+        );
+
+        $auth = new BearerAuthentication(
+            new TestValidator()
+        );
+
+        $tokenInfo = $auth->isAuthenticated($request);
+        $this->assertTrue($tokenInfo->get('active'));
+        $this->assertSame('fkooman', $tokenInfo->get('sub'));
+    }
+
+    public function testIsAuthenticatedValidQueryParameter()
+    {
+        $request = new Request(
+            array(
+                'SERVER_NAME' => 'www.example.org',
+                'SERVER_PORT' => 80,
+                'QUERY_STRING' => 'access_token=t_fkooman',
+                'REQUEST_URI' => '/?access_token=t_fkooman',
+                'SCRIPT_NAME' => '/index.php',
+                'REQUEST_METHOD' => 'GET',
             )
         );
 
@@ -68,6 +91,25 @@ class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
         $this->assertFalse($auth->isAuthenticated($request));
     }
 
+    public function testIsAuthenticatedInvalidQueryParameter()
+    {
+        $request = new Request(
+            array(
+                'SERVER_NAME' => 'www.example.org',
+                'SERVER_PORT' => 80,
+                'QUERY_STRING' => 'access_token=xyz',
+                'REQUEST_URI' => '/?access_token=xyz',
+                'SCRIPT_NAME' => '/index.php',
+                'REQUEST_METHOD' => 'GET',
+            )
+        );
+
+        $auth = new BearerAuthentication(
+            new TestValidator()
+        );
+        $this->assertFalse($auth->isAuthenticated($request));
+    }
+
     public function testIsAuthenticatedNoAttempt()
     {
         $request = new Request(
@@ -87,10 +129,6 @@ class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
         $this->assertFalse($auth->isAuthenticated($request));
     }
 
-    /**
-     * @expectedException fkooman\Http\Exception\UnauthorizedException
-     * @expectedExceptionMessage no_token
-     */
     public function testRequestAuthenticationNoToken()
     {
         $request = new Request(
@@ -104,16 +142,30 @@ class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
             )
         );
         $auth = new BearerAuthentication(
-            new TestValidator()
+            new TestValidator(),
+            array(
+                'my_var' => 'foo',
+            )
         );
 
-        $auth->requestAuthentication($request);
+        try {
+            $auth->requestAuthentication($request);
+            $this->assertTrue(false);
+        } catch (UnauthorizedException $e) {
+            $this->assertSame(
+                array(
+                    'HTTP/1.1 401 Unauthorized',
+                    'Content-Type: application/json',
+                    'Content-Length: 20',
+                    'Www-Authenticate: Bearer my_var="foo",realm="Protected Resource"',
+                    '',
+                    '{"error":"no_token"}',
+                ),
+                $e->getJsonResponse()->toArray()
+            );
+        }
     }
 
-    /**
-     * @expectedException fkooman\Http\Exception\UnauthorizedException
-     * @expectedExceptionMessage invalid_token
-     */
     public function testRequestAuthenticationInvalidToken()
     {
         $request = new Request(
@@ -128,9 +180,27 @@ class BearerAuthenticationTest extends PHPUnit_Framework_TestCase
             )
         );
         $auth = new BearerAuthentication(
-            new TestValidator()
+            new TestValidator(),
+            array(
+                'my_var' => 'foo',
+            )
         );
 
-        $auth->requestAuthentication($request);
+        try {
+            $auth->requestAuthentication($request);
+            $this->assertTrue(false);
+        } catch (UnauthorizedException $e) {
+            $this->assertSame(
+                array(
+                    'HTTP/1.1 401 Unauthorized',
+                    'Content-Type: application/json',
+                    'Content-Length: 25',
+                    'Www-Authenticate: Bearer my_var="foo",realm="Protected Resource",error="invalid_token"',
+                    '',
+                    '{"error":"invalid_token"}',
+                ),
+                $e->getJsonResponse()->toArray()
+            );
+        }
     }
 }